remove some of the alias logic
The aliases (configurable through the `~/.docker/config.json`) were a temporary
solution while buildx was not yet the default. They're an internal-use (undocumented)
option, so we don't have to take other uses into account.

This removes some of the logic to read the aliasses-config from the configuration
file, and hard-codes it to assume `buildx`.

(Probably more code can be removed/simplified)

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>